chore: move context functions into their own file
